Move around code for httpResolver
